@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 #  flutter_datetime_picker: ^1.5.0
 #    backdrop: ^0.5.3
 #    intl: ^0.15.7
@@ -29,18 +29,20 @@ dependencies:
     sdk: flutter
   flutter_datetime_picker: ^1.5.0
   backdrop: ^0.5.3
-  intl: ^0.15.7
+  intl: ^0.17.0
   http: ^0.12.2
   flutter_easyloading: ^2.2.2
   html: ^0.15.0
   flutter_html:
+
 #  flutter_view_widgets: ^1.0.6
 #  flutter_webview_plugin: ^0.3.11
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-
+  flutter_localizations: # Add this line
+    sdk: flutter
 dev_dependencies:
   flutter_test:
     sdk: flutter
@@ -55,7 +57,7 @@ flutter:
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
-
+  #generate: true # Add this line
   assets:
     - assets/
   # An image asset can refer to one or more resolution-specific "variants", see
@@ -83,3 +85,5 @@ flutter:
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
+flutter_intl:
+  enabled: true
