@@ -1,10 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:backdrop/backdrop.dart';
-import 'file:///E:/AndroidStudioProjects/flutter_app_than_so_hoc_2/lib/Pages/introl/intro.dart';
+import 'package:flutter_app_than_so_hoc_2/Pages/setting/setting.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:intl/intl.dart';
+import 'Pages/introl/intro.dart';
+
 
 // @dart=2.9
 import 'Pages/home/home.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
+
+import 'generated/l10n.dart';
 
 void main() => runApp(new MyApp());
 
@@ -19,18 +25,26 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   @override
+  AppLocalizationDelegate delegate_load = S.delegate;
   void initState() {
     // TODO: implement initState
     super.initState();
   }
 
   int _currentIndex = 0;
-  final List<Widget> _pages = [HomePage(), introPage()];
+  final List<Widget> _pages = [HomePage(), introPage(),settingPage()];
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'THẦN SỐ HỌC',
+      title: "THAN SO HOC",
+      localizationsDelegates: [
+        delegate_load,
+        GlobalMaterialLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+        GlobalCupertinoLocalizations.delegate,
+      ],
+      supportedLocales: S.delegate.supportedLocales,
       initialRoute: '/',
       builder: EasyLoading.init(),
       home: BackdropScaffold(
@@ -42,7 +56,8 @@ class _MyAppState extends State<MyApp> {
             topLeft: Radius.circular(50), topRight: Radius.circular(50)),
         appBar: BackdropAppBar(
           title: Text(
-            "THẦN SỐ HỌC",
+              // ignore: unnecessary_null_comparison
+              S != null ? S.of(context).home : "Home 2",
             // style: TextStyle(color: Colors.black),
           ),
           backgroundColor: Color(0xffc9a70e),
@@ -67,6 +82,11 @@ class _MyAppState extends State<MyApp> {
               "Giới thiệu",
               style: TextStyle(color: Colors.white),
             )),
+            ListTile(
+                title: Text(
+                  "EN",
+                  style: TextStyle(color: Colors.white),
+                )),
           ],
           onTap: (int position) => {setState(() => _currentIndex = position)},
         ),
